@@ -4,12 +4,12 @@ import 'package:flutter/material.dart';
 
 ///A controller for updating the single digit displayed by [DigitTicker]
 class _DigitTickerController extends ValueNotifier {
-  int _number;
+  _DigitTickerController({int num = 0})
+      : assert(num <= 9 && num >= 0, 'Number must be in the range of [0, 10).'),
+        _number = num,
+        super(num);
 
-  _DigitTickerController({int num = 0}) : super(num) {
-    assert(num <= 9 && num >= 0, 'Number must be in the range of [0, 10).');
-    this._number = num;
-  }
+  int _number;
 
   int get number => _number;
 
@@ -42,17 +42,15 @@ class _DigitTicker extends StatelessWidget {
   ///The controller of this widget.
   final _DigitTickerController controller;
 
-  _DigitTicker(
-      {this.backgroundColor = Colors.transparent,
-      this.curve = Curves.ease,
-      this.controller,
-      this.textStyle,
-      this.initialNumber,
-      this.duration})
-      : scrollController = ScrollController(
-            initialScrollOffset: textStyle.fontSize * (4 / 3) * initialNumber),
-        assert(controller != null, 'Controller cannot be null.'),
-        assert(initialNumber != null, 'Initial number cannot be null.'),
+  _DigitTicker({
+    this.backgroundColor = Colors.transparent,
+    this.curve = Curves.ease,
+    this.textStyle = const TextStyle(color: Colors.black, fontSize: 12),
+    this.duration = const Duration(milliseconds: 300),
+    required this.controller,
+    required this.initialNumber,
+  })  : scrollController = ScrollController(
+            initialScrollOffset: textStyle.fontSize! * (4 / 3) * initialNumber),
         assert(initialNumber <= 9),
         assert(initialNumber >= 0),
         super(key: Key(controller.toString())) {
@@ -63,7 +61,7 @@ class _DigitTicker extends StatelessWidget {
   void onValueChanged() {
     if (this.scrollController.hasClients) {
       scrollController.animateTo(
-          controller.number * textStyle.fontSize * (4 / 3),
+          controller.number * textStyle.fontSize! * (4 / 3),
           duration: duration,
           curve: curve);
     }
@@ -73,8 +71,8 @@ class _DigitTicker extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
         color: backgroundColor,
-        height: textStyle.fontSize * (4 / 3),
-        width: textStyle.fontSize * (7 / 10),
+        height: textStyle.fontSize! * (4 / 3),
+        width: textStyle.fontSize! * (7 / 10),
         child: Stack(
           children: [
             ListView(
@@ -83,7 +81,7 @@ class _DigitTicker extends StatelessWidget {
               children: [
                 for (var i in List.generate(10, (index) => index))
                   Container(
-                    height: textStyle.fontSize * (4 / 3),
+                    height: textStyle.fontSize! * (4 / 3),
                     child: Center(
                       child: Text("$i", style: textStyle),
                     ),
@@ -96,11 +94,11 @@ class _DigitTicker extends StatelessWidget {
 }
 
 class NumberTickerController extends ValueNotifier {
-  double _number;
+  NumberTickerController()
+      : _number = 0,
+        super(0);
 
-  NumberTickerController() : super(0) {
-    this._number = 0;
-  }
+  double _number;
 
   double get number => _number;
 
@@ -136,14 +134,11 @@ class NumberTicker extends StatefulWidget {
   NumberTicker(
       {this.backgroundColor = Colors.transparent,
       this.curve = Curves.ease,
-      this.controller,
+      required this.controller,
       this.textStyle = const TextStyle(color: Colors.black, fontSize: 12),
-      this.initialNumber,
+      required this.initialNumber,
       this.duration = const Duration(milliseconds: 300),
-      this.fractionDigits = 0})
-      : assert(controller != null, 'Controller cannot be null.'),
-        assert(initialNumber != null, 'Initial number cannot be null.'),
-        assert(duration != null, 'Duration cannot be null.') {
+      this.fractionDigits = 0}) {
     this.controller.number = initialNumber;
   }
 
@@ -155,22 +150,22 @@ class NumberTicker extends StatefulWidget {
 class _NumberTickerState extends State<NumberTicker>
     with SingleTickerProviderStateMixin {
   ///The animation controller for animating the removed or added [DigitTicker].
-  AnimationController animationController;
+  late AnimationController animationController;
 
   ///The number this widget currently displaying.
-  double currentNum;
+  late double currentNum;
 
   ///The string representation of the [currentNum].
-  String currentNumString;
+  late String currentNumString;
 
   ///The index of decimal point in the [currentNum].
-  int decimalIndex;
+  late int decimalIndex;
 
   ///The indicator of whether the new number is longer or shorter than the current one.
   bool shorter = false, longer = false;
 
   ///The list of [_DigitTickerController].
-  List<_DigitTickerController> digitControllers = [];
+  List<_DigitTickerController?> digitControllers = [];
 
   @override
   void initState() {
@@ -243,7 +238,7 @@ class _NumberTickerState extends State<NumberTicker>
           : currentNumString.codeUnitAt(i) - 48;
 
       if (digit >= 0 && digit != oldDigit) {
-        digitControllers[i].number = digit;
+        digitControllers[i]?.number = digit;
       } else {}
     }
 
@@ -259,7 +254,7 @@ class _NumberTickerState extends State<NumberTicker>
 
   @override
   Widget build(BuildContext context) {
-    var width = widget.textStyle.fontSize * (7 / 10);
+    var width = widget.textStyle.fontSize! * (7 / 10);
     return AnimatedBuilder(
         animation: animationController,
         builder: (_, __) {
@@ -275,7 +270,7 @@ class _NumberTickerState extends State<NumberTicker>
                       ? Text(' ')
                       : _DigitTicker(
                           backgroundColor: widget.backgroundColor,
-                          controller: digitControllers.first,
+                          controller: digitControllers.first!,
                           curve: widget.curve,
                           duration: widget.duration,
                           textStyle: widget.textStyle,
@@ -286,7 +281,7 @@ class _NumberTickerState extends State<NumberTicker>
                       ? Text(' ')
                       : _DigitTicker(
                           backgroundColor: widget.backgroundColor,
-                          controller: digitControllers[i],
+                          controller: digitControllers[i]!,
                           curve: widget.curve,
                           duration: widget.duration,
                           textStyle: widget.textStyle,
@@ -302,7 +297,7 @@ class _NumberTickerState extends State<NumberTicker>
                         ? Text(' ')
                         : _DigitTicker(
                             backgroundColor: widget.backgroundColor,
-                            controller: digitControllers.last,
+                            controller: digitControllers.last!,
                             curve: widget.curve,
                             duration: widget.duration,
                             textStyle: widget.textStyle,
